libc/fildes: Define all instances of file-tx ops next to their users
Merge all the _tx_ops.{c,h} files into their respective _tx.c files.
The content of the _ops files is the vtable of the latter and the latter
are the only callers. Both belong in the same source file.
