--- modules/libc/src/fildes/dir_tx_ops.h
@@ -1,32 +1,0 @@
-/*
- * picotm - A system-level transaction manager
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- *
- * SPDX-License-Identifier: LGPL-3.0-or-later
- */
-
-#pragma once
-
-#include "file_tx_ops.h"
-
-/**
- * \cond impl || libc_impl || libc_impl_fd
- * \ingroup libc_impl
- * \ingroup libc_impl_fd
- * \file
- * \endcond
- */
-
-extern const struct file_tx_ops dir_tx_ops;
